Move SessionHandle classes to Parser namespace
